Refactor: factories and tests, delete related gfc objects
1. Factories moved form test folder to app level
2. Delete method extended to delete objects with Generic Relations
3. Test for deleting generic related objects
