+++ apps/articles/factories/__init__.py
@@ -1,0 +1,3 @@
+from .blog_items import BlogItemContentModuleFactory, BlogItemFactory
+from .news_items import NewsItemContentModuleFactory, NewsItemFactory
+from .projects import ProjectContentModuleFactory, ProjectFactory
