+++ apps/articles/tests/conftest.py
@@ -1,0 +1,32 @@
+import pytest
+
+from apps.core.factories import ImageFactory, PersonFactory
+from apps.info.factories import FestivalFactory
+from apps.library.factories import PerformanceFactory, PlayFactory
+
+pytestmark = [pytest.mark.django_db]
+
+
+@pytest.fixture
+def images():
+    ImageFactory.create_batch(10)
+
+
+@pytest.fixture
+def festivals(images):
+    return FestivalFactory.create_batch(5)
+
+
+@pytest.fixture
+def plays(festivals):
+    return PlayFactory.create_batch(10)
+
+
+@pytest.fixture
+def persons():
+    return PersonFactory.create_batch(10)
+
+
+@pytest.fixture
+def performances(persons, plays):
+    return PerformanceFactory.complex_create(3)
